# server1/swagger.json
{
    "openapi":"3.0.0",
    "info":{
        "title":"API de manpulação do Motor de Vendas",
        "description":"Essa API tem como objetivo a manipulação do Motor de Vendas",
        "version":"1.0.0"
    },
    "servers":[
        {
            "url":"http://localhost:3007",
            "description":"API de teste"
        }
    ],
    "paths":{
...
        "/client":{
            "post":{
                "summary":"Cadastro  de cliente",
                "description":"Essa rota é responsável pelo cadastro de um novo cliente",
                "tags":["Cliente"],
                "security":[{"bearerAuth":[]}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "required":["name","cnpj","razaoSocial","dateClient"],
                                
                                "properties": {
                                    "name":{
                                        "type": "string",
                                        "example":"Malwee",
                                        "description": "Nome do Cliente"
                                    },
                                    "cnpj":{
                                        "type": "string",
                                        "example":"12345678901234",
                                        "description": "cnpj do cliente"
                                    },
                                    "razaoSocial":{
                                        "type":"string",
                                        "example":"Malwee",
                                        "description":"Razão Social da empresa"

                                    },
                                    "dateClient":{
                                        "type":"string",
                                        "format":"date",
                                        "example":"2017-07-21T17:32:28Z",
                                        "description":"Cliente desde"

                                    },
                                    "adddress":{
                                        "type":"array",
                                        "items":{
                                            "type":"object",
                                            "properties":{
                                                "rua":{
                                                    "type":"string",
                                                    "example":"Rua Bertha Weege",
                                                    "description":"rua do cliente"
                                                },
                                                "bairro":{
                                                    "type":"string",
                                                    "example":"Barra do Rio Cerro",
                                                    "description":"bairro do cliente"
                                                },
                                                "cidade":{
                                                    "type":"string",
                                                    "example":"Jaraguá do Sul",
                                                    "description":"cidade do cliente"
                                                },
                                                "estado":{
                                                    "type":"string",
                                                    "example":"SC",
                                                    "description":"estado do cliente"
                                                },
                                                "cep":{
                                                    "type":"string",
                                                    "example":"89260-900",
                                                    "description":"cep do cliente"  
                                                },
                                                "numero":{
                                                    "type":"integer",
                                                    "example":200,
                                                    "description":"numero do cliente"  
                                                },
                                                "complemento":{
                                                    "type":"string",
                                                    "example":"Rua principal",
                                                    "description":"complemento"
                                                },
                                                "pontoDeReferencia":{
                                                    "type":"string",
                                                    "example":"parque Malwee",
                                                    "description":"complemento"
                                                }
                                            }

                                        },
                                        "description":"Array de endereços"

                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"

                    },
                    "401": {
                        "description":"product already exists!"
                    },
                    "403":{
                        "description":"Invalid token"

                    },
                    "200":{
                        "description": "ok"
                        
                    }
                }
                
            },
            "get":{
                "description":"Buscar todos clientes",
                "summary":"Essa rota é responsável por buscar todos clientes",
                "tags":["Cliente"],
                "security":[{"bearerAuth":[]}],
                "parameters": [],
                "responses": {
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref": "#/components/schemas/Clients"
                                    }
                                    
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }
            },
            "put":{
                "description":"Edita um produto",
                "summary":"Essa rota é responsável por editar um produto passando seu id pelo body",
                "tags":["Cliente"],
                "security":[{"bearerAuth":[]}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "required":["name","cnpj","razaoSocial","dateClient"],
                                
                                "properties": {
                                    "name":{
                                        "type": "string",
                                        "example":"Malwee",
                                        "description": "Nome do Cliente"
                                    },
                                    "razaoSocial":{
                                        "type":"string",
                                        "example":"Malwee",
                                        "description":"Razão Social da empresa"

                                    },
                                    "idClient":{
                                        "type":"integer",
                                        "example":1,
                                        "description":"id do cliente"

                                    },
                                    "adddress":{
                                        "type":"array",
                                        "items":{
                                            "type":"object",
                                            "properties":{
                                                "rua":{
                                                    "type":"string",
                                                    "example":"Rua Bertha Weege",
                                                    "description":"rua do cliente"
                                                },
                                                "bairro":{
                                                    "type":"string",
                                                    "example":"Barra do Rio Cerro",
                                                    "description":"bairro do cliente"
                                                },
                                                "cidade":{
                                                    "type":"string",
                                                    "example":"Jaraguá do Sul",
                                                    "description":"cidade do cliente"
                                                },
                                                "estado":{
                                                    "type":"string",
                                                    "example":"SC",
                                                    "description":"estado do cliente"
                                                },
                                                "cep":{
                                                    "type":"string",
                                                    "example":"89260-900",
                                                    "description":"cep do cliente"  
                                                },
                                                "numero":{
                                                    "type":"integer",
                                                    "example":200,
                                                    "description":"numero do cliente"  
                                                },
                                                "complemento":{
                                                    "type":"string",
                                                    "example":"Rua principal",
                                                    "description":"complemento"
                                                },
                                                "pontoDeReferencia":{
                                                    "type":"string",
                                                    "example":"parque Malwee",
                                                    "description":"complemento"
                                                },
                                                "id":{
                                                    "type":"integer",
                                                    "description":"Id do endereço",
                                                    "example":1
                                                }
                                            }

                                        },
                                        "description":"Array de endereços"

                                    }
                                }
                            }
                        }
                    }
                },

                "responses":{
                    "200":{
                        "description":"OK"
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }
            }
        },
        "/client/{id}":{
            "get":{
                "description": "Buscar um único cliente pelo id",
                "summary": "Essa rota é responsável por buscar um único cliente pelo id",
                "tags": ["Cliente"],
                "security":[{"bearerAuth":[]}],
                "parameters": [{
                        "name":"id",
                        "in":"path",
                        "description": "Id do cliente",
                        "required": true
                    }
                ],
                "responses": {
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref": "#/components/schemas/Clients"
                                    }
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }
            },
            "patch":{
                "description": "Edita o 'status' de um cliente para '0' pelo id",
                "summary": "Essa rota é responsável por inativar o cliente pelo id",
                "tags": ["Cliente"],
                "security":[{"bearerAuth":[]}],
                "parameters": [{
                        "name":"id",
                        "in":"path",
                        "description": "Id do cliente",
                        "required": true
                    }
                ],
                "responses": {
                    "400":{
                        "description":"Request needs a token"
                    },
                    "200":{
                        "description": "OK"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"

                    },
                    "403":{
                        "description":"Invalid token"

                    }

                }
            },
            "post":{
                "summary":"Cadastro  de cliente",
                "description":"Essa rota é responsável pelo cadastro de um novo cliente",
                "tags":["Cliente"],
                "security":[{"bearerAuth":[]}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "required":["name","cnpj","razaoSocial","dateClient"],
                                
                                "properties": {
                                    "name":{
                                        "type": "string",
                                        "example":"Malwee",
                                        "description": "Nome do Cliente"
                                    },
                                    "razaoSocial":{
                                        "type":"string",
                                        "example":"Malwee",
                                        "description":"Razão Social da empresa"

                                    },
                                    "idClient":{
                                        "type":"integer",
                                        "example":1,
                                        "description":"id do cliente"

                                    },
                                    "adddress":{
                                        "type":"array",
                                        "items":{
                                            "type":"object",
                                            "properties":{
                                                "rua":{
                                                    "type":"string",
                                                    "example":"Rua Bertha Weege",
                                                    "description":"rua do cliente"
                                                },
                                                "bairro":{
                                                    "type":"string",
                                                    "example":"Barra do Rio Cerro",
                                                    "description":"bairro do cliente"
                                                },
                                                "cidade":{
                                                    "type":"string",
                                                    "example":"Jaraguá do Sul",
                                                    "description":"cidade do cliente"
                                                },
                                                "estado":{
                                                    "type":"string",
                                                    "example":"SC",
                                                    "description":"estado do cliente"
                                                },
                                                "cep":{
                                                    "type":"string",
                                                    "example":"89260-900",
                                                    "description":"cep do cliente"  
                                                },
                                                "numero":{
                                                    "type":"integer",
                                                    "example":200,
                                                    "description":"numero do cliente"  
                                                },
                                                "complemento":{
                                                    "type":"string",
                                                    "example":"Rua principal",
                                                    "description":"complemento"
                                                },
                                                "pontoDeReferencia":{
                                                    "type":"string",
                                                    "example":"parque Malwee",
                                                    "description":"complemento"
                                                },
                                                "id":{
                                                    "type":"integer",
                                                    "description":"Id do endereço",
                                                    "example":1
                                                }
                                            }

                                        },
                                        "description":"Array de endereços"

                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"

                    },
                    "401": {
                        "description":"product already exists!"
                    },
                    "403":{
                        "description":"Invalid token"

                    },
                    "200":{
                        "description": "ok"
                        
                    }
                }
            }
        },
        "/ClientsEndereco/{id}":{
            "get":{
                "description": "Buscar um único endereco do cliente pelo id",
                "summary": "Essa rota é responsável por buscar um único endereco do cliente pelo id",
                "tags": ["Cliente"],
                "security":[{"bearerAuth":[]}],
                "parameters": [{
                        "name":"id",
                        "in":"path",
                        "description": "Id do endereço",
                        "required": true
                    }
                ],
                "responses": {
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref": "#/components/schemas/Enderecos"
                                    }
                                }
                            }
                        }
                    },
                    "400":{
                        "description":"Request needs a token"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"
    
                    },
                    "403":{
                        "description":"Invalid token"
    
                    }
                }
            },
            "patch":{
                "description": "Edita o 'status' de um endereco do cliente para '0' pelo id",
                "summary": "Essa rota é responsável por inativar um endereco do cliente pelo id",
                "tags": ["Cliente"],
                "security":[{"bearerAuth":[]}],
                "parameters": [{
                        "name":"id",
                        "in":"path",
                        "description": "Id do endereco",
                        "required": true
                    }
                ],
                "responses": {
                    "400":{
                        "description":"Request needs a token"
                    },
                    "200":{
                        "description": "OK"
                    },
                    "404":{
                        "description":"Token invalid. Its necessary bearer + token"

                    },
                    "403":{
                        "description":"Invalid token"

                    }

                }
            }
        },
...
    },
    "components":{
        "schemas":{
...
            "Clients":{
                "required":["name","cnpj","razaoSocial","dateClient"],
                "properties":{
                    "name":{
                        "type":"string",
                        "description":"Nome do Cliente"
                    },
                    "cnpj":{
                        "type":"string",
                        "decription":"cnpj da empresa"
                    },
                    "razaoSocial":{
                        "type":"string",
                        "description":"Razão Social da empresa"
                    },
                    "dateClient":{
                        "type":"string",
                        "format":"date",
                        "example":"2017-07-21T17:32:28Z",
                        "description":"Cliente desde"
                    }
                }
            },
            "Enderecos":{
                "properties":{
                    "rua":{
                        "type":"string",
                        "description":"rua do cliente"
                    },
                    "bairro":{
                        "type":"string",
                        "description":"bairro do cliente"
                    },
                    "cidade":{
                        "type":"string",
                        "description":"cidade do cliente"
                    },
                    "estado":{
                        "type":"string",
                        "description":"estado do cliente"
                    },
                    "cep":{
                        "type":"string",
                        "description":"cep do cliente"  
                    },
                    "numero":{
                        "type":"integer",
                        "description":"numero do cliente"  
                    },
                    "complemento":{
                        "type":"string",
                        "description":"complemento"
                    },
                    "pontoDeReferencia":{
                        "type":"string",
                        "description":"complemento"
                    },
                    "fkClients":{
                        "type":"integer",
                        "description":"id do cliente"  
                    }
                }
            },
...
        },
        "securitySchemes":{
            "bearerAuth":{
                "type":"http",
                "scheme":"bearer",
                "bearerFormat":"JWT"
            }
        }
    }
}